+++ smartcontract/service/native/init/init.go
@@ -20,11 +20,11 @@ package init
 
 import (
 	"github.com/ontio/dad-go/smartcontract/service/native/auth"
+	"github.com/ontio/dad-go/smartcontract/service/native/fee_split"
 	params "github.com/ontio/dad-go/smartcontract/service/native/global_params"
+	"github.com/ontio/dad-go/smartcontract/service/native/governance"
 	"github.com/ontio/dad-go/smartcontract/service/native/ong"
 	"github.com/ontio/dad-go/smartcontract/service/native/ont"
-	"github.com/ontio/dad-go/smartcontract/service/native/governance"
-	"github.com/ontio/dad-go/smartcontract/service/native/feeSplit"
 	"github.com/ontio/dad-go/smartcontract/service/native/oracle"
 )
 
@@ -35,5 +35,5 @@ func init() {
 	auth.Init()
 	oracle.InitOracle()
 	governance.InitGovernance()
-	feeSplit.InitFeeSplit()
+	fee_split.InitFeeSplit()
 }
